♻️PostsManagerPage 분리 완료
