+++ src/features/post/api/index.ts
@@ -1,0 +1,4 @@
+export * from "./FetchPosts"
+export * from "./FetchPostsByTags"
+export * from "./FetchTags"
+export * from "./SearchPosts"
