+++ src/shared/ui/index.tsx
@@ -6,3 +6,4 @@ export * from "./Input.tsx"
 export * from "./Select.tsx"
 export * from "./Table.tsx"
 export * from "./TextArea.tsx"
+export * from "./Pagination.tsx"
