+++ src/widgets/ui/AddCommentDialog.tsx
@@ -1,0 +1,47 @@
+import { useStore } from "../../app/store"
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../shared/ui/Dialog"
+import { Textarea, Button } from "../../shared/ui"
+
+/**
+ *  댓글 추가 대화상자
+ * @returns
+ */
+
+export const AddCommentDialog = () => {
+  const { showAddCommentDialog, newComment, setShowAddCommentDialog, setComments, setNewComment } = useStore()
+
+  // 댓글 추가
+  const addComment = async () => {
+    try {
+      const response = await fetch("/api/comments/add", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(newComment),
+      })
+      const data = await response.json()
+      setComments(data.postId, data)
+      setShowAddCommentDialog(false)
+      setNewComment({ body: "", postId: null, userId: 1 })
+    } catch (error) {
+      console.error("댓글 추가 오류:", error)
+    }
+  }
+
+  return (
+    <Dialog open={showAddCommentDialog} onOpenChange={setShowAddCommentDialog}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>새 댓글 추가</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <Textarea
+            placeholder="댓글 내용"
+            value={newComment.body}
+            onChange={(e) => setNewComment({ ...newComment, body: e.target.value })}
+          />
+          <Button onClick={addComment}>댓글 추가</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
