+++ src/widgets/ui/AddPostDialog.tsx
@@ -2,32 +2,32 @@ import React from "react"
 import { useStore } from "../../app/store"
 import { Dialog, DialogContent, DialogHeader, DialogTitle, Input, Textarea, Button } from "../../shared/ui"
 
-// 상태 관리
-const { posts, newPost, showAddDialog, setPosts, setNewPost, setShowAddDialog } = useStore()
-
-// 게시물 추가
-const addPost = async () => {
-  try {
-    const response = await fetch("/api/posts/add", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify(newPost),
-    })
-    const data = await response.json()
-    setPosts([data, ...posts])
-    setShowAddDialog(false)
-    setNewPost({ id: 0, title: "", body: "", userId: 1 })
-  } catch (error) {
-    console.error("게시물 추가 오류:", error)
-  }
-}
-
 /**
  * 게시물 추가 대화상자
  * @returns
  */
 
 export const AddPostDialog = () => {
+  // 상태 관리
+  const { posts, newPost, showAddDialog, setPosts, setNewPost, setShowAddDialog } = useStore()
+
+  // 게시물 추가
+  const addPost = async () => {
+    try {
+      const response = await fetch("/api/posts/add", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(newPost),
+      })
+      const data = await response.json()
+      setPosts([data, ...posts])
+      setShowAddDialog(false)
+      setNewPost({ id: 0, title: "", body: "", userId: 1 })
+    } catch (error) {
+      console.error("게시물 추가 오류:", error)
+    }
+  }
+
   return (
     <Dialog open={showAddDialog} onOpenChange={setShowAddDialog}>
       <DialogContent>
