+++ src/widgets/ui/index.tsx
@@ -1,5 +1,11 @@
 export * from "./AddPostDialog"
+export * from "./EditPostDialog"
+export * from "./AddCommentDialog"
+export * from "./EditCommentDialog"
+export * from "./ViewPostDialog"
 export * from "./Footer"
 export * from "./Header"
 export * from "./PostTable"
 export * from "./Comments"
+export * from "./ControlContainer"
+export * from "./UserModalDialog"
